EOD 2/26/2016
Working on getting dummyStorageOutflowCurves.py to work. Integrated scripts run otherwise, but tables aren't written because Optimatics code assumes the tables exist when it edits them. Need to create the tables. Current error: Can't load IA 32-bit .dll on a AMD 64-bit platform. Probably need to get a 32-bit version of java and connect it to the project.